# docs/swagger-doc.yaml
openapi: 3.0.1
info:
  title: Tunemyles API
  description: >-
    This is a sample server Petstore server. You can find out more about Swagger
    at [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net](irc.freenode.net)
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'http://localhost:5000/api/v2'
    description: Dev Server API URL
...
paths:
...
  '/masters/banks/{id}/account':
    patch:
      tags:
        - masters
      summary: Update Bank Detail (Admin Route)
      description: Update Bank Detail (Admin Route)
      operationId: updateBankDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBankAccountBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
  '/masters/banks/{id}/logo':
    patch:
      tags:
        - masters
      summary: Change Bank Logo (Admin Route)
      description: Change Bank Logo (Admin Route)
      operationId: changeBankLogo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateBankLogoBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  '/masters/categories/{id}':
    patch:
      tags:
        - masters
      summary: Update Category (Admin Route)
      description: Update Category (Admin Route)
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
    delete:
      tags:
        - masters
      summary: Delete Category (Admin Route)
      description: Delete Category (Admin Route)
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  '/masters/categories/{id}/icon':
    patch:
      tags:
        - masters
      summary: Change Category Icon (Admin Route)
      description: Change Category Icon (Admin Route)
      operationId: changeCategoryIcon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChangeCategoryIconBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  '/masters/medias/{id}':
    patch:
      tags:
        - masters
      summary: Update Media (Admin Route)
      description: Update Meida (Admin Route)
      operationId: updateMedia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostMediaBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
    delete:
      tags:
        - masters
      summary: Delete Media (Admin Route)
      description: Delete Media (Admin Route)
      operationId: deleteMedia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  '/masters/faqs/{id}':
    patch:
      tags:
        - masters
      summary: Update Faq (Admin Route)
      description: Update Faq (Admin Route)
      operationId: updateFaq
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFaqBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
    delete:
      tags:
        - masters
      summary: Delete Faq (Admin Route)
      description: Delete FAq (Admin Route)
      operationId: deleteFaq
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: number
        success:
          type: boolean
        messages:
          type: string
...
    UpdateBankAccountBody:
      required:
        - account_number
        - account_name
      type: object
      properties:
        account_number:
          type: string
        account_name:
          type: string
    UpdateBankLogoBody:
      required:
        - logo
      properties:
        logo:
          type: string
          format: binary
...
    UpdateCategoryBody:
      type: object
      properties:
        category:
          type: string
        slug:
          type: string
        description:
          type: string
        visibility:
          type: integer
    ChangeCategoryIconBody:
      required:
        - icon
      properties:
        icon:
          type: string
          format: binary
    PostMediaBody:
      required:
        - image
      properties:
        image:
          type: string
          format: binary
...
    UpdateFaqBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      name: authorization
      in: header